Beginning work on a Map implementation that uses a series of rules to dynamically restrict which keys are accessible.
The restrictions are lazily evaluated, so the first accessor call after a restriction is added may take a while
The LinkedRestrictionMap's operations that are inherited from Map run in O(1) time after the first call, which also evaluates the restrictions.
Applying a restriction after the restrictions have been computed is an O(1) operation as well.